fix compile/linkage errors on Mac
git-svn-id: https://svn.code.sf.net/p/oorexx/code-0/interpreter-3.x/trunk@2052 0b6cbdbe-3aab-466e-b73a-abd511dda0a2

+++ api/platform/unix/rexxplatformapis.h
@@ -58,9 +58,9 @@ APIRET APIENTRY RexxPullQueue (
         const char *,                          /* Name of queue to read from  */
         PRXSTRING,                             /* RXSTRING to receive data    */
         REXXDATETIME *,                        /* Stor for data date/time     */
-        unsigned int);                         /* wait status (WAIT|NOWAIT)   */
+        size_t);                               /* wait status (WAIT|NOWAIT)   */
 typedef APIRET (APIENTRY *PFNREXXPULLQUEUE)(const char *, PCONSTRXSTRING, REXXDATETIME *,
-                                           unsigned int);
+                                           size_t);
 
 #endif /* REXXPLATFORMAPIS_INCLUDED */
